Refactor Ui components as per PR feedback
