--- src/main/java/seedu/address/ui/ReminderPersonListPanel.java
@@ -1,49 +1,0 @@
-package seedu.address.ui;
-
-import java.util.logging.Logger;
-
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.scene.control.ListCell;
-import javafx.scene.control.ListView;
-import javafx.scene.layout.Region;
-import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.person.Person;
-
-/**
- * Panel containing the list of persons with Reminders.
- */
-public class ReminderPersonListPanel extends UiPart<Region> {
-    private static final String FXML = "ReminderPersonListPanel.fxml";
-    private final Logger logger = LogsCenter.getLogger(PersonListPanel.class);
-
-    @FXML
-    private ListView<Person> reminderPersonListView;
-
-    /**
-     * Creates a {@code PersonListPanel} with the given {@code ObservableList}.
-     */
-    public ReminderPersonListPanel(ObservableList<Person> personList) {
-        super(FXML);
-        reminderPersonListView.setItems(personList);
-        reminderPersonListView.setCellFactory(listView -> new ReminderPersonListViewCell());
-    }
-
-    /**
-     * Custom {@code ListCell} that displays the graphics of a {@code Person} using a {@code PersonCard}.
-     */
-    class ReminderPersonListViewCell extends ListCell<Person> {
-        @Override
-        protected void updateItem(Person person, boolean empty) {
-            super.updateItem(person, empty);
-
-            if (empty || person == null) {
-                setGraphic(null);
-                setText(null);
-            } else {
-                setGraphic(new ReminderPersonCard(person, getIndex() + 1).getRoot());
-            }
-        }
-    }
-
-}
